Feature: Add courses app
You can add a course and add several lessons to each course, and you can specify whether the lesson is free or not. You can also add an evaluation to the course if the student is only registered in it. You can add comments and respond to comments in any lesson.